# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Prediction Game API",
    "description": "API for wallet-authenticated prediction gameplay, leaderboards, rounds, and predictions. Use Swagger UI to explore endpoints and test requests.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste a JWT like: Bearer <token>"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseErrorResponse"
          },
          {
            "type": "object",
            "properties": {
              "details": {
                "type": "array",
                "description": "Field-level validation details (present on validation errors only)",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string",
                      "example": "walletAddress"
                    },
                    "message": {
                      "type": "string",
                      "example": "walletAddress is required"
                    }
                  },
                  "required": [
                    "field",
                    "message"
                  ]
                }
              }
            }
          }
        ]
      },
      "BaseErrorResponse": {
        "type": "object",
        "description": "Standard error response returned by all API endpoints on failure.",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error class name (e.g. ValidationError, AuthenticationError, NotFoundError)",
            "example": "ValidationError"
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the error",
            "example": "walletAddress is required"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code for programmatic handling",
            "example": "VALIDATION_ERROR"
          }
        },
        "required": [
          "error",
          "message",
          "code"
        ]
      },
      "MoneyAmount": {
        "type": "string",
        "description": "Canonical monetary amount: a Decimal(20,8) serialized as a string with exactly 8 fractional digits. Never a JSON number — IEEE-754 floats cannot represent stroop-scale values safely.",
        "pattern": "^-?\\d+\\.\\d{8}$",
        "example": "1000.33333333"
      },
...
      "RateLimitResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        ],
        "example": {
          "error": "AppError",
          "message": "Too many requests from this IP, please try again after 15 minutes",
          "code": "RATE_LIMIT_EXCEEDED"
        }
      },
...
      "LeaderboardEntry": {
        "type": "object",
        "properties": {
          "rank": {
            "type": "integer"
          },
          "userId": {
            "type": "string"
          },
          "walletAddress": {
            "type": "string"
          },
          "totalEarnings": {
            "$ref": "#/components/schemas/MoneyAmount"
          },
          "totalPredictions": {
            "type": "integer"
          },
          "accuracy": {
            "type": "number"
          },
          "modeStats": {
            "type": "object",
            "properties": {
              "upDown": {
                "type": "object",
                "properties": {
                  "wins": {
                    "type": "integer"
                  },
                  "losses": {
                    "type": "integer"
                  },
                  "earnings": {
                    "$ref": "#/components/schemas/MoneyAmount"
                  },
                  "accuracy": {
                    "type": "number"
                  }
                }
              },
              "legends": {
                "type": "object",
                "properties": {
                  "wins": {
                    "type": "integer"
                  },
                  "losses": {
                    "type": "integer"
                  },
                  "earnings": {
                    "$ref": "#/components/schemas/MoneyAmount"
                  },
                  "accuracy": {
                    "type": "number"
                  }
                }
              }
            }
          }
        }
      },
      "LeaderboardResponse": {
        "type": "object",
        "properties": {
          "leaderboard": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LeaderboardEntry"
            }
          },
          "userPosition": {
            "$ref": "#/components/schemas/LeaderboardEntry",
            "nullable": true
          },
          "totalUsers": {
            "type": "number"
          },
          "lastUpdated": {
            "type": "string"
          }
        },
        "required": [
          "leaderboard",
          "totalUsers",
          "lastUpdated"
        ],
        "additionalProperties": true
      },
...
    }
  },
...
  "paths": {
...
    "/api/leaderboard": {
      "get": {
        "summary": "Get the global leaderboard",
        "tags": [
          "leaderboard"
        ],
        "responses": {
          "200": {
            "description": "Leaderboard payload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  },
                  "$ref": "#/components/schemas/LeaderboardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "description": "Returns the global leaderboard. Bearer authentication is **optional**; if provided,\nthe API includes the requesting user's position.\n\nSupports two pagination modes:\n\n**Cursor mode** (recommended for large datasets): pass `cursor` from a previous\nresponse's `pagination.nextCursor` to load the next page. Efficient for deep pages.\n\n**Offset mode**: pass `offset` to skip rows. Backward-compatible with existing clients.\n\nWhen `cursor` is present, `offset` is ignored.\n",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            },
            "description": "Max number of entries to return (max 500)"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Pagination offset (offset mode, ignored when cursor is present)"
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            },
            "description": "Opaque cursor from pagination.nextCursor (cursor mode)"
          }
        ],
        "x-codeSamples": [
          {
            "lang": "cURL",
            "source": "curl -X GET \"$API_BASE_URL/api/leaderboard?limit=100&offset=0\"\n"
          }
        ]
      }
    },
    "/api/leaderboard/batch": {
      "post": {
        "summary": "Get leaderboard positions for multiple users",
        "description": "Retrieve leaderboard positions and stats for multiple users in a single request.\nEach user lookup is processed independently. Partial success is supported.\nMaximum 100 user IDs per request.\n",
        "tags": [
          "leaderboard"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "minItems": 1,
                    "maxItems": 100,
                    "description": "Array of user IDs to query"
                  }
                },
                "required": [
                  "userIds"
                ]
              },
              "example": {
                "userIds": [
                  "user-id-1",
                  "user-id-2",
                  "user-id-3"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch user positions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "true if at least one query succeeded"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BatchUserPositionResult"
                      }
                    }
                  }
                },
                "example": {
                  "success": true,
                  "results": [
                    {
                      "userId": "user-id-1",
                      "position": {
                        "rank": 15,
                        "userId": "user-id-1",
                        "walletAddress": "GBRPY...4B",
                        "totalEarnings": 125.5,
                        "totalPredictions": 42,
                        "accuracy": 73.81,
                        "modeStats": {
                          "upDown": {
                            "wins": 20,
                            "losses": 10,
                            "earnings": 85.25,
                            "accuracy": 66.67
                          },
                          "legends": {
                            "wins": 12,
                            "losses": 8,
                            "earnings": 40.25,
                            "accuracy": 60
                          }
                        }
                      }
                    },
                    {
                      "userId": "user-id-2",
                      "error": "User not found"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "examples": {
                  "emptyUserIds": {
                    "value": {
                      "error": "At least one user ID is required"
                    }
                  },
                  "tooManyUserIds": {
                    "value": {
                      "error": "Maximum 100 user IDs per query"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "No token provided"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Failed to fetch batch user positions"
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "cURL",
            "source": "curl -X POST \"$API_BASE_URL/api/leaderboard/batch\" \\\\\n  -H \"Content-Type: application/json\" \\\\\n  -H \"Authorization: Bearer $TOKEN\" \\\\\n  -d '{\"userIds\":[\"user-id-1\",\"user-id-2\"]}'\n"
          }
        ]
      }
    },
...
    "/api/chat/send": {
      "post": {
        "tags": [
          "Chat"
        ],
        "summary": "Send a chat message",
        "description": "Authenticated users only. Rate limit: **5 messages per minute per user**. On limit, responds with **429**.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "content": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message created"
          },
          "429": {
            "description": "Too many messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/history": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Get chat history",
        "description": "Returns recent chat messages. Supports two pagination modes:\n\n**Cursor mode** (recommended for real-time chat): pass `cursor` from a\nprevious response to load the next (older) page. Messages are returned\noldest-first within each page.\n\n**Offset mode**: pass `offset` to skip rows. Backward-compatible with\nexisting clients that only pass `limit`.\n\nWhen `cursor` is present, offset is ignored.\n",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 50
            },
            "description": "Number of messages per page (max 50)"
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            },
            "description": "Opaque cursor from pagination.nextCursor (cursor mode)"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Number of rows to skip (offset mode, ignored when cursor is present)"
          }
        ],
        "responses": {
          "200": {
            "description": "Chat history page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatHistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
  }
}